HW-3 network package, UpbitApi interface 생성
