+++ NackUn/StudyFirst/app/src/main/java/sample/nackun/com/studyfirst/network/UpbitApi.kt
@@ -1,0 +1,15 @@
+package sample.nackun.com.studyfirst.network
+
+import retrofit2.Call
+import retrofit2.http.GET
+import retrofit2.http.Query
+import sample.nackun.com.studyfirst.vo.Market
+import sample.nackun.com.studyfirst.vo.Ticker
+
+interface UpbitApi {
+    @GET("market/all")
+    fun requestMarket(): Call<ArrayList<Market>>
+
+    @GET("ticker/")
+    fun requestTicker(@Query("markets") markets: String): Call<ArrayList<Ticker>>
+}
